requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.1
PyYAML==6.0
mysql-connector-python==8.1.0
lxml==4.9.3